+++ core/pbwowstyle.php
@@ -204,7 +204,6 @@ class pbwowstyle
 		// Misc
 		$tpl_vars += array(
 			'HEADERLINKS_CODE' 	    => ($headerlinks_enable && $headerlinks_code) ? str_replace('&', '&amp;', html_entity_decode($headerlinks_code)) : false,
-			'ADS_INDEX_CODE' 	    => ($ads_index_enable && $ads_index_code) ? str_replace('&', '&amp;', html_entity_decode($ads_index_code)) : false,
 		);
 
 		// Assign vars
@@ -212,6 +211,32 @@ class pbwowstyle
 		$this->template->append_var('BODY_CLASS', $body_class);
 	}
 
+
+	/**
+	 * Returns the formatted advertisement block HTML for the index page, or false if disabled.
+	 * Called by the avathar.recenttopicsav.modify_ads_code event listener.
+	 *
+	 * @return string|false
+	 */
+	public function get_ads_index_code()
+	{
+		$pbwow_config = $this->pbwow_config;
+
+		if (!isset($pbwow_config) || !is_array($pbwow_config))
+		{
+			return false;
+		}
+
+		$ads_index_enable = $pbwow_config['ads_index_enable'] ?? false;
+		$ads_index_code = $pbwow_config['ads_index_code'] ?? '';
+
+		if ($ads_index_enable && $ads_index_code)
+		{
+			return str_replace('&', '&amp;', html_entity_decode($ads_index_code));
+		}
+
+		return false;
+	}
 
 	/**
 	 * Gets the PBWoW config data from the DB, or the cache if it is present